Handle missing git version info...
Which also requires a 'make-dist' script, since I am too lazy
to do this properly in cmake. Really. That said, there is
a 'make dist', but you don't wan't to use it.

The versioning is tricky. The root is the fact that a file
containing the git hash cannot be committed, since the hash isn't
known until after the commit is made.

This setup supports two usecases:
  - Building from a git repo, obtaining version from git.
  - Building from a tarball created from make-dist, which contains the
    git version info found when creating the tarball.

It should degrade gracefully if neiher of these assumptions holds.